+++ MantleArrayExtension/Classes/MAEArrayAdapter.m
@@ -562,8 +562,7 @@ static unichar const MAEDefaultSeparator = ' ';
 {
     NSParameterAssert(objCType != nil);
 
-    if (strcmp(objCType, @encode(char)) == 0
-        || strcmp(objCType, @encode(int)) == 0
+    if (strcmp(objCType, @encode(int)) == 0
         || strcmp(objCType, @encode(short)) == 0
         || strcmp(objCType, @encode(long)) == 0
         || strcmp(objCType, @encode(long long)) == 0
@@ -575,8 +574,12 @@ static unichar const MAEDefaultSeparator = ' ';
         || strcmp(objCType, @encode(float)) == 0
         || strcmp(objCType, @encode(double)) == 0) {
         return [self.class numberTransformer];
-    } else if (strcmp(objCType, @encode(BOOL)) == 0) {
+    } else if (strcmp(objCType, @encode(BOOL)) == 0
+               || strcmp(objCType, @encode(bool)) == 0) {
         return [self.class boolTransformer];
+    } else if (strcmp(objCType, @encode(char)) == 0) {
+        // NOTE: In the case of 32 bit OS, BOOL equals to char, so it only applies to other cases.
+        return [self.class numberTransformer];
     }
 
     return nil;