+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - It possible to easily specify expecting that it equals to specified characters.
   - Please refer to a class that name is `MAERawFragment`.
+- Add `MAEArrayAdapter # valueByFragmentWithFormat:separatedStrings:error:`
+  - Implement of `classForParsingArray` will easier.
 
 ### Major changes